refactor : validator 패키지와 Utils 클래스를 utils 패키지로 통합
