+++ src/main/java/racingcar/utils/StringSeparator.java
@@ -1,12 +1,12 @@
-package racingcar;
+package racingcar.utils;
 
 import java.util.Arrays;
 
-public class Utils {
-    public static final String STRING_REGEX = ",";
+public class StringSeparator {
+    public static final String COMMA = ",";
 
     public static String[] splitAndTrim(String names) {
-        String[] splitNames = names.split(STRING_REGEX);
+        String[] splitNames = names.split(COMMA);
         Arrays.setAll(splitNames, i -> splitNames[i].trim());
         return splitNames;
     }
