+++ src/main/java/racingcar/view/InputView.java
@@ -1,12 +1,12 @@
 package racingcar.view;
 
-import racingcar.validator.MoveCountValidator;
-import racingcar.validator.NameValidator;
+import racingcar.utils.MoveCountValidator;
+import racingcar.utils.NameValidator;
 
 import java.util.Scanner;
 
 public class InputView {
-    public static final String REQUEST_CAR_NAMES = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
+    private static final String REQUEST_CAR_NAMES = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
     private static final String REQUEST_MOVE_COUNT = "시도할 회수는 몇회인가요?";
 
     public static String getCarNames(Scanner scanner) {
