+++ src/test/java/racingcar/MoveCountValidatorTest.java
@@ -3,10 +3,9 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.validator.MoveCountValidator;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static racingcar.validator.MoveCountValidator.*;
+import static racingcar.utils.MoveCountValidator.*;
 
 class MoveCountValidatorTest {
     @ParameterizedTest
